# perf: ~3x faster server side `_(translate)` calls