--- a/frappe/tests/test_translate.py
+++ b/frappe/tests/test_translate.py
@@ -29,11 +29,13 @@ class TestTranslate(unittest.TestCase):
 	def setUp(self):
 		if self._testMethodName in self.guest_sessions_required:
 			frappe.set_user("Guest")
+		frappe.local.lang_full_dict = None  # reset cached translations
 
 	def tearDown(self):
 		frappe.form_dict.pop("_lang", None)
 		if self._testMethodName in self.guest_sessions_required:
 			frappe.set_user("Administrator")
+		frappe.local.lang_full_dict = None  # reset cached translations
 
 	def test_extract_message_from_file(self):
 		data = frappe.translate.get_messages_from_file(translation_string_file)