Local caching of the merged translation dictionary was broken because it was attempting to find language code in the dictionary, which makes no sense because:
1. Language can't change during the same request. 
2. Dictionary is eng->translated string, why would it have language code in there?

before
```
In [1]: %timeit frappe._("Sales Invoice", lang="de")
7.54 µs ± 14.6 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

after
```
In [1]: %timeit frappe._("Sales Invoice", lang="de")
2.09 µs ± 7.11 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```